Server/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
bs4==0.0.1
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
chuck-norris-python==1.2
clarifai-grpc==6.7.1
click==7.1.2
commonmark==0.9.1
cryptography==3.1
decorator==4.4.2
dictdiffer==0.8.1
Flask==1.1.2
googleapis-common-protos==1.52.0
grpcio==1.31.0
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
notion==0.0.25
paho-mqtt==1.5.0
protobuf==3.13.0
pycparser==2.20
pyee==7.0.2
pyppeteer==0.2.2
pyppeteer-stealth==1.0.2
python-slugify==4.0.1
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
text-unidecode==1.3
tornado==6.0.4
tqdm==4.48.2
tzlocal==2.1
urllib3==1.25.10
validators==0.17.1
webscraping==1.7.1
websockets==8.1
Werkzeug==1.0.1
